rename attendances to sessions https://trello.com/c/6y3yxPKO
