+++ meteor-app/imports/ui/containers/role-register-container.js
@@ -2,7 +2,7 @@ import { Meteor } from 'meteor/meteor'
 import { withTracker } from 'meteor/react-meteor-data';
 import RoleRegister from '../components/roles-main/role-register'
 import Roles from '/imports/api/roles/roles';
-import Attendances from '/imports/api/attendances/attendances';
+import Sessions from '/imports/api/sessions/sessions';
 
 export default withTracker((props) => {
   const rolesHandle = Meteor.subscribe('all.roles')
